requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
python-socketio==5.10.0
python-engineio==4.8.0
pandas==2.1.4
numpy==1.26.2
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.16.2
chromedriver-autoinstaller==0.6.4
webdriver-manager==4.0.1
openai==1.6.1
langdetect==1.0.9
wikipedia==1.4.0
pyfiglet==1.0.2
google-api-python-client==2.111.0
openpyxl==3.1.2
eventlet==0.33.3
psutil==5.9.6
Pillow==10.1.0
lxml==4.9.3
urllib3==2.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6